+++ podcasts/ChaptersViewController.swift
@@ -30,6 +30,7 @@ class ChaptersViewController: PlayerItemViewController {
 
     override func willBeAddedToPlayer() {
         updateColors()
+        header.update()
         addObservers()
     }
 
@@ -65,6 +66,7 @@ class ChaptersViewController: PlayerItemViewController {
     @objc private func update() {
         chaptersTable.reloadData()
         updateColors()
+        header.update()
     }
 
     private func updateColors() {